feat(ops): ropeAxes takes fractional positions
Issue #200. `positions` was `Int32Array` because Z-Image indexes tokens by
their grid coordinate. MiniMax-H3's visual VAE does not: it normalises each
axis to `(-1, 1)` -- `2 * (i + 0.5) / n - 1` -- and multiplies the angle by
`2*pi`, so its positions are fractional.

A rotation by a fractional angle is the same rotation. Nothing in the
arithmetic ever needed a whole number, and the alternative was a second kernel
differing from this one by a single binding type -- two conventions waiting to
drift. `f32` is the general one and every integer a model here uses is exact in
it; a negative position still stays negative and turns the rotation the other
way, which is the property `i32` was chosen for.

The reference takes `Float32Array | Int32Array`, so existing callers are
unchanged. The one GPU caller uploads f32.

**A mutation survived, and the gap was real.** Putting `f32(i32(pos))` back
into the kernel -- exactly the change that undoes this -- left every case in
`wgsl-axes.test.ts` green, because all of them use Z-Image's whole-number
positions. The capability was in the binding and in no test.
`h3-axes.wgsl.test.ts` is one dispatch with fractional positions, and both that
mutation and `trunc(pos)` now go red.

`h3-axes.test.ts` pins the whole *mapping* onto H3's rope rather than only the
capability, because a rope applied to the wrong channel pairs returns a tensor
of the right shape whose norm is unchanged:

  - the **frequencies already agree**. H3 builds `theta ** -arange(0, 1,
    2*n_dim/rot_dim)` -- eight exponents for `rot_dim = 48, n_dim = 3` -- and
    `ropeAxes` builds `theta ** -(2*pair/axis_dim)` over a 16-channel axis,
    which is the same eight. Nothing had to change.
  - the **pairing does not**. H3 tiles `[t8, h8, w8]` twice and rotates halves;
    `ropeAxes` pairs adjacent channels inside each axis block.
    `H3_ROPE_PERMUTATION` converts one to the other and is applied to the
    *weights*, once, exactly as `permuteForRope` does for Anima.
  - **`rope_dim_ratio: 0.75`** leaves 16 of 64 channels unrotated. A fourth axis
    pinned at position 0 is the identity, so `axisDims [16,16,16,16]` covers it
    with no second code path.

Worst element 2.384e-7, which is f32 rounding on values of order one. Four
mutations against the mapping go red: the permutation dropped, the idle axis
given a real position, the axis split made uneven, the frequency normalised by
the head instead of the axis.

The fixture is generated by `ops/rope/tools/gen_h3_fixture.py`, which
**transcribes** H3's three functions rather than importing them -- a weaker
guarantee than `gen_axes_fixture.py`'s, and the generator says so: H3 ships its
code inside the checkpoint as remote code, and constructing
`RotaryEmbeddingND` pulls in diffusers, distributed helpers and a
flash-attention import.

One assertion was written carelessly and the fixture was right: "every position
is fractional" is false, because `2 * (i + 0.5) / n - 1` is exactly 0 at the
middle of an odd-length axis. It asserts the structure now -- eight whole
positions, all zero, one per `T * W` token -- so a regenerated fixture with
different dims fails and is looked at rather than sliding under a threshold.

1731 tests pass.

Refs #200

+++ ops/rope/wgsl-axes.test.ts
@@ -24,7 +24,7 @@ const code = kernel(import.meta.url, "axes");
 const ZIMAGE = { axisDims: [32, 48, 48], thetaBase: 256, N: 5, numHeads: 3 } as const;
 const zimage = axesScenario(code, {
   ...ZIMAGE,
-  positions: Int32Array.from([0, 0, 0, 1, 0, 1, 2, 3, 5, 7, 1, 0, 3, 11, 4]),
+  positions: Float32Array.from([0, 0, 0, 1, 0, 1, 2, 3, 5, 7, 1, 0, 3, 11, 4]),
 });
 
 // Unequal axes, and negative positions on two of them. Both halves are
@@ -37,7 +37,7 @@ const uneven = axesScenario(code, {
   thetaBase: 256,
   N: 9,
   numHeads: 8,
-  positions: Int32Array.from(
+  positions: Float32Array.from(
     Array.from({ length: 9 }, (_, token) => [token, -3 * token, 40 - token]).flat(),
   ),
 });
@@ -51,7 +51,7 @@ const ONE = { axisDims: [16], thetaBase: 10000, N: 9, numHeads: 8 } as const;
 const oneAxisInput = Float32Array.from({ length: 9 * 8 * 16 }, (_, i) => Math.sin(i * 0.29) * 2);
 const oneAxis = axesScenario(
   code,
-  { ...ONE, positions: Int32Array.from({ length: 9 }, (_, token) => token + 200) },
+  { ...ONE, positions: Float32Array.from({ length: 9 }, (_, token) => token + 200) },
   oneAxisInput,
   rope({ input: oneAxisInput, N: 9, numHeads: 8, headDim: 16, posOffset: 200, thetaBase: 10000 }),
 );
@@ -65,7 +65,7 @@ const oneAxis = axesScenario(
 const still = Float32Array.from({ length: 5 * 3 * 128 }, (_, i) => Math.sin(i * 0.29) * 2 + 3);
 const identity = axesScenario(
   code,
-  { ...ZIMAGE, positions: new Int32Array(5 * 3) },
+  { ...ZIMAGE, positions: new Float32Array(5 * 3) },
   still,
 );
